+++ packages/widget-core/src/widgets/base/config.ts
@@ -1,0 +1,6 @@
+export interface BaseConfig {
+    floatieInstance?: string
+    sendButtonLabel?: string
+    confirmationMessage?: string
+    size?: 'medium' | 'large'
+  }