+++ packages/widget-core/src/widgets/basic/config.ts
@@ -1,0 +1,8 @@
+import type { BaseConfig } from '../base/config'
+
+export interface BasicWidgetConfig extends BaseConfig {
+  offsetBottom?: number
+  offsetRight?: number
+  title?: string
+  placeholder?: string
+}