+++ packages/widget-core/src/widgets/basic/data.ts
@@ -1,0 +1,4 @@
+export interface BasicWidgetData {
+  type: 'BUG' | 'FEATURE' | 'COMMENT'
+  message: string
+}