+++ packages/widget-core/src/widgets/index.ts
@@ -1,0 +1,7 @@
+export * from './basic'
+export * from './rating'
+
+export enum WidgetType {
+  Basic = 'Basic',
+  Rating = 'Rating',
+}