+++ packages/widget-core/src/widgets/rating/config.ts
@@ -1,0 +1,5 @@
+import type { BaseConfig } from '../base/config'
+
+export interface RatingWidgetConfig extends BaseConfig {
+  range: number
+}