+++ packages/widget-nuxt/playground/nuxt.config.ts
@@ -1,4 +1,6 @@
 export default defineNuxtConfig({
   modules: ['../src/module'],
-  devtools: { enabled: true },
+  floatie: {
+    clientKey: 'poop',
+  },
 })