+++ packages/widget-nuxt/src/runtime/components/FloatieBasic.vue
@@ -1,0 +1,17 @@
+<script lang="ts" setup>
+import { FloatieBasic as Basic } from '@floatie/widget-vue'
+import type { BasicWidgetConfig } from '@floatie/widget-core'
+
+const props = defineProps<{ config?: BasicWidgetConfig }>()
+
+const {
+  public: { floatie },
+} = useRuntimeConfig()
+</script>
+
+<template>
+  <Basic
+    :config="props.config"
+    :client-key="floatie.clientKey"
+  />
+</template>