+++ packages/widget-vue/src/index.ts
@@ -1,3 +1,4 @@
-import Widget from './components/Widget.vue'
+import FloatieBasic from './components/Basic.vue'
+import FloatieRating from './components/Rating.vue'
 
-export { Widget }
+export { FloatieBasic, FloatieRating }