+++ packages/widget-core/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @floatie/widget-core
 
+## 3.1.0
+
+### Minor Changes
+
+- Refactor the way different types of widgets are built
+
 ## 3.0.0
 
 ### Major Changes