+++ packages/widget-vue/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @floatie/widget-vue
 
+## 3.1.0
+
+### Minor Changes
+
+- Refactor the way different types of widgets are built
+
+### Patch Changes
+
+- Updated dependencies
+  - @floatie/widget-core@3.1.0
+
 ## 3.0.0
 
 ### Major Changes